rename: useGetHalfway -> calcHalfwayPoint 이름변경, utils/place 폴더 경로 이동
